backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.15
deprecation==2.1.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
frozenlist==1.5.0
gotrue==2.11.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
inflect==7.4.0
lru_cache==0.2.3
Mako==1.3.8
MarkupSafe==3.0.2
more-itertools==10.5.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
postgrest==0.18.0
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
realtime==2.0.6
requests==2.32.3
rsa==4.9
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
storage3==0.9.0
supafunc==0.7.0
svix==1.44.0
typeguard==4.4.1
types-Deprecated==1.2.15.20241117
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==13.1
wrapt==1.17.0
yarl==1.18.3
